--- src/main/java/darren/assistant/DarrenAssistant.java
@@ -1,9 +1,0 @@
-package darren.assistant;
-
-public class DarrenAssistant {
-    public static void main(String[] args) {
-        new DarrenJr().run();
-    }
-}
-
-
