--- src/main/java/darren/assistant/ui/DarrenAssistantException.java
@@ -1,7 +1,0 @@
-package darren.assistant.ui;
-
-public class DarrenAssistantException extends Exception{
-    public DarrenAssistantException(String message) {
-        super(message);
-    }
-}
