+++ src/main/java/java/ui/DarrenAssistantException.java
@@ -1,0 +1,7 @@
+package ui;
+
+public class DarrenAssistantException extends Exception{
+    public DarrenAssistantException(String message) {
+        super(message);
+    }
+}
